// workspaces/litellm-orchestrator/plugins/litellm-orchestrator/src/schema/openapi/generated/models/UserKey.model.ts
//

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************

/**
 * A user key
 * @public
 */
export interface UserKey {
  userId: string;
  keyId: string;
  keySecret: string;
}